*Really* fixed short-summary issue and...
Turned mpaa if/elsif nest into a case statement which reads slightly better.  Used a pretty ugly lambda to fix the issue with short summaries not being shown. Tried single-line if/else logic that didn't work. Could certainly do multiline if/else, but in the interest of space, I think not for now. Could be implemented in the CLI instead (as could mpaa).