Added walkthrough. Audio issues! Use VLC/mPlayer!
Apologize for the audio quality. It's a laptop, so the mic is built-in right above the fans and my voice is sometime pretty hard to understand and a bit distorted. QuickTime doesn't play this back very well, it's more audible in VLC and you can of course crank the volume up past 100% there, so that's a good bet.